apps/ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdfplumber==0.11.1
PyMuPDF==1.24.5
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.31
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9
openpyxl==3.1.5
reportlab==4.2.2
